bits: impl bit masking ops for Mask
