+++ bits/tests/mask.rs
@@ -1,6 +1,7 @@
+use bits::Mask;
+
 #[test]
 fn a_and_b() {
-    use bits::And;
     let v1: &[u8] = &[0b_1111_0000, 0b_0000_1111, 0b_1010_1010];
     let v2: &[u8] = &[0b_0011_1100, 0b_0011_1100, 0b_0101_0101];
     let v3: &[u8] = &[0b_1100_0011, 0b_1100_0011, 0b_1010_1010];
@@ -14,7 +15,6 @@ fn a_and_b() {
 
 #[test]
 fn a_not_b() {
-    use bits::Not;
     let v1: &[u8] = &[0b_1111_1111, 0b_1111_1111, 0b_1111_1111];
     let v2: &[u8] = &[0b_0000_1111, 0b_1111_0000, 0b_0101_0101];
     for (index, bits) in v1.not(v2) {
@@ -24,7 +24,6 @@ fn a_not_b() {
 
 #[test]
 fn a_or_b() {
-    use bits::Or;
     let v1: &[u8] = &[0b_1111_0000, 0b_0000_1111, 0b_1010_1010];
     let v2: &[u8] = &[0b_0000_1111, 0b_1111_0000, 0b_0101_0101];
     for (_index, bits) in v1.or(v2) {
@@ -34,7 +33,6 @@ fn a_or_b() {
 
 #[test]
 fn a_xor_b() {
-    use bits::Xor;
     let v1: &[u8] = &[0b_1111_0000, 0b_0000_1111, 0b_1010_1010];
     let v2: &[u8] = &[0b_0011_0011, 0b_1100_1100, 0b_0110_1001];
     for (_index, bits) in v1.xor(v2) {
